+++ lib/crunchbase/utilities/entity_endpoints.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative '../errors'
+
 module Crunchbase
   # Utilities
   module Utilities
@@ -15,8 +17,8 @@ module Crunchbase
       #   limit:      Number of rows to return. Default is 100, min is 1, max is 100.
 
       # Lookup an Organization or single card
-      def organization(entity_id, card_id: nil)
-        lookup_for('organization', entity_id, card_id)
+      def organization(entity_id, **card_args)
+        lookup_for('organization', entity_id, card_args)
       end
 
       # Lookup Organization's all cards
@@ -25,8 +27,8 @@ module Crunchbase
       end
 
       # Lookup a Person or single card
-      def person(entity_id, card_id: nil)
-        lookup_for('person', entity_id, card_id)
+      def person(entity_id, **card_args)
+        lookup_for('person', entity_id, card_args)
       end
 
       # Lookup Person's all cards
@@ -35,8 +37,8 @@ module Crunchbase
       end
 
       # Lookup a Funding Round or single card
-      def funding_round(entity_id, card_id: nil)
-        lookup_for('funding_round', entity_id, card_id)
+      def funding_round(entity_id, **card_args)
+        lookup_for('funding_round', entity_id, card_args)
       end
 
       # Lookup Funding Round's all cards
@@ -45,8 +47,8 @@ module Crunchbase
       end
 
       # Lookup an Acquisition or Single card
-      def acquisition(entity_id, card_id: nil)
-        lookup_for('acquisition', entity_id, card_id)
+      def acquisition(entity_id, **card_args)
+        lookup_for('acquisition', entity_id, card_args)
       end
 
       # Lookup Acquisition's all card
@@ -55,8 +57,8 @@ module Crunchbase
       end
 
       # Lookup an Investment or Single card
-      def investment(entity_id, card_id: nil)
-        lookup_for('investment', entity_id, card_id)
+      def investment(entity_id, **card_args)
+        lookup_for('investment', entity_id, card_args)
       end
 
       # Lookup Investment's all card
@@ -65,8 +67,8 @@ module Crunchbase
       end
 
       # Lookup an PressReference or Single card
-      def press_reference(entity_id, card_id: nil)
-        lookup_for('press_reference', entity_id, card_id)
+      def press_reference(entity_id, **card_args)
+        lookup_for('press_reference', entity_id, card_args)
       end
 
       # Lookup PressReference's all card
@@ -75,8 +77,8 @@ module Crunchbase
       end
 
       # Lookup an Ipo or Single card
-      def ipo(entity_id, card_id: nil)
-        lookup_for('ipo', entity_id, card_id)
+      def ipo(entity_id, **card_args)
+        lookup_for('ipo', entity_id, card_args)
       end
 
       # Lookup Ipo's all card
@@ -85,8 +87,8 @@ module Crunchbase
       end
 
       # Lookup an fund or Single card
-      def fund(entity_id, card_id: nil)
-        lookup_for('fund', entity_id, card_id)
+      def fund(entity_id, **card_args)
+        lookup_for('fund', entity_id, card_args)
       end
 
       # Lookup fund's all card
@@ -95,8 +97,8 @@ module Crunchbase
       end
 
       # Lookup an fund or Single card
-      def ownership(entity_id, card_id: nil)
-        lookup_for('ownership', entity_id, card_id)
+      def ownership(entity_id, **card_args)
+        lookup_for('ownership', entity_id, card_args)
       end
 
       # Lookup fund's all card
@@ -110,11 +112,13 @@ module Crunchbase
         Crunchbase::Entities::Client.new(entity_id, entity_type)
       end
 
-      def lookup_for(entity_type, entity_id, card_id)
+      def lookup_for(entity_type, entity_id, **card_args)
         kobject = entities(entity_type, entity_id)
+
+        card_id = card_args&.delete(:card_id)
         return kobject.fetch if card_id.nil?
 
-        kobject.cards(card_id)
+        kobject.cards(card_id, card_args)
       end
     end
   end